move Modal directly into WorkDetails
fixes weird issue where checkbox was closing modal, but not sure why
